Lab_Azure_Functions/Azure_Function_App/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-projects==1.0.0b2
azure-storage-queue
azure-identity